LayoutTests/inspector/css/get-system-fonts-expected.txt

This test ensures that the inspector can enumerate system font families, and checks for the existence of common fonts.

Includes Arial
Includes Times
Missing DejaVu Sans
Missing Liberation Serif
